+++ service/src/main/kotlin/io/provenance/explorer/domain/models/explorer/TransactionModels.kt
@@ -71,7 +71,8 @@ data class TxDetails(
     val signers: Signatures,
     val memo: String,
     val monikers: Map<String, String>,
-    val feepayer: TxFeepayer
+    val feepayer: TxFeepayer,
+    var additionalHeights: List<Int> = emptyList()
 )
 
 data class TxFeepayer(val type: String, val address: String)