+++ service/src/main/kotlin/io/provenance/explorer/service/AccountService.kt
@@ -44,6 +44,7 @@ import io.provenance.explorer.domain.models.explorer.TokenCounts
 import io.provenance.explorer.domain.models.explorer.UnpaginatedDelegation
 import io.provenance.explorer.domain.models.explorer.mapToProtoCoin
 import io.provenance.explorer.domain.models.explorer.toCoinStrWithPrice
+import io.provenance.explorer.domain.models.explorer.toData
 import io.provenance.explorer.grpc.extensions.getModuleAccName
 import io.provenance.explorer.grpc.extensions.isStandardAddress
 import io.provenance.explorer.grpc.extensions.isVesting
@@ -151,6 +152,11 @@ class AccountService(
                 bals.count().toLong()
             )
         }
+    }
+
+    // Used for balance validation - this is raw data only
+    fun getAccountBalancesAllAtHeight(address: String, height: Int) = runBlocking {
+        accountClient.getAccountBalancesAllAtHeight(address, height).map { it.toData() }
     }
 
     private fun getDelegationTotal(address: String) = runBlocking {